Fix CPU tick overflow: change tick counters from u32 to u64
Linux /proc/stat counters are effectively 64-bit and can exceed u32::MAX
on long-running systems. Parsing into u32 causes silent overflow before
wraparound logic can handle it correctly. macOS Mach API values fit in
u32 but promoting to u64 is safe and maintains consistency.

+++ src/sampler/linux.rs
@@ -118,14 +118,14 @@ impl LinuxKernel {
                 let parts: Vec<&str> = line.split_whitespace().collect();
                 // Format: cpuN user nice system idle iowait irq softirq [steal [guest [guest_nice]]]
                 if parts.len() >= 5 {
-                    let user = parts[1].parse::<u32>().unwrap_or(0);
-                    let nice = parts[2].parse::<u32>().unwrap_or(0);
-                    let system = parts[3].parse::<u32>().unwrap_or(0);
-                    let idle = parts[4].parse::<u32>().unwrap_or(0);
-                    let iowait = parts.get(5).and_then(|s| s.parse().ok()).unwrap_or(0u32);
-                    let irq = parts.get(6).and_then(|s| s.parse().ok()).unwrap_or(0u32);
-                    let softirq = parts.get(7).and_then(|s| s.parse().ok()).unwrap_or(0u32);
-                    let steal = parts.get(8).and_then(|s| s.parse().ok()).unwrap_or(0u32);
+                    let user = parts[1].parse::<u64>().unwrap_or(0);
+                    let nice = parts[2].parse::<u64>().unwrap_or(0);
+                    let system = parts[3].parse::<u64>().unwrap_or(0);
+                    let idle = parts[4].parse::<u64>().unwrap_or(0);
+                    let iowait = parts.get(5).and_then(|s| s.parse().ok()).unwrap_or(0u64);
+                    let irq = parts.get(6).and_then(|s| s.parse().ok()).unwrap_or(0u64);
+                    let softirq = parts.get(7).and_then(|s| s.parse().ok()).unwrap_or(0u64);
+                    let steal = parts.get(8).and_then(|s| s.parse().ok()).unwrap_or(0u64);
 
                     cpu_ticks.push(CpuTicks {
                         user,
